+++ test/helpers/wait.js
@@ -1,0 +1,9 @@
+function wait(ms) { 
+  return new Promise((resolve, reject) => {
+    setTimeout(resolve, ms);
+  });
+};
+
+module.exports = {
+  wait
+};
